refactor: 불필요한 Generator 인터페이스 도출 삭제
